Move Draw* traits to sub-module
